requirements.txt
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
charset-normalizer==3.4.6
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.1.0
langchain-core==1.2.20
langgraph==1.1.3
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.12
langsmith==0.7.22
openai==2.29.0
orjson==3.11.7
ormsgpack==1.12.2
packaging==23.2
pydantic==2.12.5
pydantic_core==2.41.5
PyYAML==6.0.1
requests==2.31.0
requests-toolbelt==1.0.0
sniffio==1.3.1
tenacity==8.5.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
xxhash==3.6.0
zstandard==0.25.0
